% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Income tax rates.}
\format{
A data frame with 6 variables:
\describe{
\item{year}{year of the income}
\item{bracket}{bracket number}
\item{min, max}{range of the bracket}
\item{rate}{rate of the bracket}
\item{currency}{always 'euros'}
}
}
\source{
\url{https://fr.wikipedia.org/wiki/Bar\%C3\%A8mes_de_l'imp\%C3\%B4t_sur_le_revenu_en_France}
}
\usage{
rates
}
\description{
A dataset containing income tax rates in France from 2006 to 2020.
}
\keyword{datasets}
